requirements.txt
# NBA Fantasy Optimizer - Python ML Dependencies

# Core ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Data Processing
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Statistical Analysis
statsmodels>=0.14.0
scipy>=1.10.0

# Database & API
mysql-connector-python>=8.0.0
sqlalchemy>=2.0.0
requests>=2.31.0

# Advanced ML & Analytics
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
scipy>=1.10.0
joblib>=1.3.0

# Time Series Analysis
prophet>=1.1.0

# Advanced ML
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Model Validation
mlxtend>=0.22.0

# Data Validation
great-expectations>=0.17.0

# Performance Monitoring
memory-profiler>=0.61.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Environment Management
python-dotenv>=1.0.0
